feat: 유저 정보 조회 및 저장 훅 추가
