+++ frontend/src/stores/authStore.ts
@@ -6,11 +6,9 @@ interface AuthActions {
     setUser: (user: User) => void
     setAuthGuest: () => void
     setAuthMember: () => void
-    setChannelId: (channelId: number) => void
 }
 
 interface AuthState {
-    channelId: number | null
     user: User | null
     isAuth: boolean
     actions: AuthActions
@@ -24,7 +22,6 @@ export const useAuthStore = create<AuthState>()(
                 user: null,
                 isAuth: false,
                 actions: {
-                    setChannelId: (channelId) => set({ channelId }),
                     setUser: (user) => set({ user }),
                     setAuthGuest: () => set({ isAuth: false }),
                     setAuthMember: () => set({ isAuth: true }),
@@ -35,7 +32,6 @@ export const useAuthStore = create<AuthState>()(
                 partialize: (state) => ({
                     user: state.user,
                     isAuth: state.isAuth,
-                    channelId: state.channelId,
                 }),
             }
         )
